% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{all_stocks}
\alias{all_stocks}
\title{Load the whole dataframe}
\usage{
all_stocks()
}
\value{
The data, as a dataframe
}
\description{
Simple function that will return the whole dataframe, as provided on Kaggle
}
\examples{
SP500_names()

}
