% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{one_stock.df}
\alias{one_stock.df}
\title{Return one stock's values}
\usage{
one_stock.df(name, var_type = "Open", start = NULL, end = NULL)
}
\arguments{
\item{name}{String value, the name of the stock. Use SP500_names for possible names}

\item{var_type}{String value, the name of the variabe to return, could be 'Open', 'Close', 'High', 'Low' or 'Volume'. Default is 'Open'}

\item{start}{Date value, the date when the serie should be started. The user should use package 'lubridate'}

\item{end}{Date value, the date when the serie should be ended. The user should use package 'lubridate'}
}
\value{
This function returns the data in a dataframe with 'Date' and 'var_type' columns.
}
\description{
Return one stock's values, as a dataframe. The user should provide the name of the stock, and possibly the type of value, and a starting and ending date
}
\examples{
one_stock.df('TSS')
one_stock.df('RMD','High',start=lubridate::ymd('2014-01-01'), end=lubridate::ymd('2015-01-01'))

}
