% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{one_stock.ts}
\alias{one_stock.ts}
\title{Return one stock's values}
\usage{
one_stock.ts(name, var_type = "Open", start = NULL, end = NULL)
}
\arguments{
\item{name}{String value, the name of the stock. Use SP500_names for possible names}

\item{var_type}{String value, the name of the variabe to return, could be 'Open', 'Close', 'High', 'Low' or 'Volume'. Default is 'Open'}

\item{start}{Date value, the date when the serie should be started. The user should use package 'lubridate'}

\item{end}{Date value, the date when the serie should be ended. The user should use package 'lubridate'}
}
\value{
This function returns the data as a time serie, with correct starting and ending time, and right frequency.
}
\description{
Return one stock's value, as a time serie. The user should provide the name of the stock, and possibly the variable, and a starting and ending date
}
\examples{
one_stock.ts('TSS')
one_stock.ts('RMD','High',start=lubridate::ymd('2014-01-01'), end=lubridate::ymd('2015-01-01'))

}
