% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SP500_names}
\alias{SP500_names}
\title{Return stocks' names}
\usage{
SP500_names()
}
\value{
List of characters, all the names available
}
\description{
Return the name of all available stocks
}
\examples{
SP500_names()

}
